Add metrix to mongo package
