+++ hmrc-mongo-metrix-play-26/src/main/scala/uk/gov/hmrc/mongo/metrix/domain/MetricSource.scala
@@ -14,16 +14,13 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.metrix.gauge
+package uk.gov.hmrc.mongo.metrix.domain
 
-import com.codahale.metrics.Gauge
-import play.api.Logger
-import uk.gov.hmrc.metrix.cache.MetricCache
+import scala.concurrent.{ExecutionContext, Future}
 
-final case class CachedMetricGauge(name: String, metrics: MetricCache) extends Gauge[Int] {
-  override def getValue: Int = {
-    val value = metrics.valueOf(name)
-    Logger.debug(s"Gauge for metric $name is reporting on value: $value")
-    value
-  }
+/**
+  * A source of metrics to collect from your application
+  */
+trait MetricSource {
+  def metrics(implicit ec: ExecutionContext): Future[Map[String, Int]]
 }
